+++ ValorDiario+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  ValorDiario+CoreDataProperties.swift
 //  SemConsumismo
 //
-//  Created by Luiza Fattori on 22/07/19.
+//  Created by Luiza Fattori on 26/07/19.
 //  Copyright © 2019 Luiza Fattori. All rights reserved.
 //
 //
@@ -17,7 +17,7 @@ extension ValorDiario {
         return NSFetchRequest<ValorDiario>(entityName: "ValorDiario")
     }
 
-    @NSManaged public var valorDinheiro: Double
-    @NSManaged public var valorTotal: Double
+    @NSManaged public var valorDinheiro: Float
+    @NSManaged public var valorTotal: Float
 
 }
